% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_fun.R
\name{debug_fun}
\alias{debug_fun}
\title{put arguments into global environment}
\usage{
debug_fun(expr)
}
\arguments{
\item{expr}{R expression}
}
\description{
put arguments into global environment
}
\examples{
arg_list <- debug_fun(expression(lm(x ~ y)))
attach(arg_list)
}
